Made a lot of changes
Removed space scene related values from ECSGame to the component; Fixed bug when exiting game in space scene it crashed the program; Moved all buttons signals and forward declarations into separate header file; fixed bug related to mouse icon; fixed bug when selected object and its sound continued even when changed scene

+++ CommonGameCode/SceneNode.cpp
@@ -193,7 +193,8 @@ void SceneNode::DeleteAllEntities()
 	for (auto child : children)
 		child->DeleteAllEntities();
 
-	ECSGame::Instance().GetEntityManager().DestroyEntity(entity);
+	if(entity.lock()!=nullptr)
+		ECSGame::Instance().GetEntityManager().DestroyEntity(entity);
 }
 
 //This function finds and returns scene node which contains this entity
